# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inazuma API",
        "description": "API for Inazuma Eleven",
        "contact": {
            "email": "gomez.sergio.smr@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "description": "Authenticates user and returns API token",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Incorrect credentials"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "description": "Revokes the current access token",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LogoutResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/coaches": {
            "get": {
                "tags": [
                    "Coaches"
                ],
                "summary": "List all coaches",
                "operationId": "59ec5701439b901b9e051a2d86df6dcf",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "name[eq]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "Seymour Hillman"
                        }
                    },
                    {
                        "name": "version[eq]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "ie1"
                        }
                    },
                    {
                        "name": "version[ne]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "ie1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CoachCollection"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Coaches"
                ],
                "summary": "Create new coach",
                "description": "Requires admin privileges",
                "operationId": "createCoach",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CoachRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Coach created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CoachResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/coaches/{id}": {
            "get": {
                "tags": [
                    "Coaches"
                ],
                "summary": "Get specific coach",
                "description": "Public endpoint to get coach details",
                "operationId": "getCoach",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Coach ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coach details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CoachResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Coach not found"
                    }
                },
                "security": []
            },
            "put": {
                "tags": [
                    "Coaches"
                ],
                "summary": "Update a coach",
                "operationId": "8e726c14ddd5577d820b8025fed37dd3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Coach"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coach updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Coach"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Coaches"
                ],
                "summary": "Delete coach",
                "description": "Requires admin privileges",
                "operationId": "deleteCoach",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Coach ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coach deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Coach deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Coach not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "1|abcdef1234567890"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            },
                            "email": {
                                "type": "string",
                                "example": "user@example.com"
                            },
                            "is_admin": {
                                "type": "boolean",
                                "example": false
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "LogoutResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Session Closed"
                    }
                },
                "type": "object"
            },
            "CoachRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Coach's name",
                        "type": "string",
                        "example": "New Coach"
                    },
                    "image": {
                        "description": "URL to coach's image",
                        "type": "string",
                        "example": "https://static.wikia.nocookie.net/inazuma/images/b/b8/Hillman_%28Sprite%29.png/revision/latest?cb=20201025114237&path-prefix=es"
                    },
                    "version": {
                        "description": "Game version",
                        "type": "string",
                        "example": "ie1"
                    }
                },
                "type": "object"
            },
            "CoachResource": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Coach"
                    }
                },
                "type": "object"
            },
            "Coach": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Seymour Hillman"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://static.wikia.nocookie.net/inazuma/images/b/b8/Hillman_%28Sprite%29.png/revision/latest?cb=20201025114237&path-prefix=es"
                    },
                    "version": {
                        "type": "string",
                        "example": "ie1"
                    }
                },
                "type": "object"
            },
            "CoachCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Coach"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/coaches?page=1"
                            },
                            "last": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/coaches?page=3"
                            },
                            "prev": {
                                "type": "string",
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/coaches?page=2"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 3
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 35
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}